move arrowicon file
